third_party/blink/web_tests/editing/deleting/password-delete-contents-expected.txt

This makes sure we are able to delete the contents of a password field: deleting a selection and evaluating the field contents.

PASS passwordField.value='helllo'; passwordField.setSelectionRange(3, 4); testRunner.execCommand('Delete', false, null); passwordField.value is 'hello'
PASS successfullyParsed is true

TEST COMPLETE
